requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
joblib==1.2.0
Levenshtein==0.20.9
lxml==4.9.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-decouple==3.7
python-Levenshtein==0.20.9
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
soupsieve==2.4
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
